Feat/master/autogeneration (#177)
* start refactor

* update services

* add task service

* tidy up nulls

* update dependencies

* Add wrapper services

* Fix namespace issues

* refactor partial #1

* refactor partial #2

* clear up  warnings and messages

* adjust tests

* adjust settings

* various fixes

* various fixes #2

* align names

* tidy up

* tidy up patterns

* slight tweaks and version bump

* handle null prefix

* update dotnet version

* update tools

* Update dotnet.yml

* Revert "Update dotnet.yml"

This reverts commit 10150f0e38294332ce08d309d6b686eac205fedd.

* update lang version

* Change dotnet version

* Restore dotnet version

* Try to specify multiple versions

* Specify 8.x dotnet version

* remove unused package

* execute task updates in paralel

* add comment for concurrent task updates

* fix usage of Type instead of WorkflowTaskType

* use type as both string and enum and adjust tests

---------

+++ src/ConductorSharp.Engine/Polling/IPollTimingStrategy.cs
@@ -7,6 +7,6 @@ namespace ConductorSharp.Engine.Polling
 {
     public interface IPollTimingStrategy
     {
-        int CalculateDelay(IDictionary<string, int> taskQueue, List<TaskToWorker> taskToWorkerList, int baseSleepInterval, int currentSleepInterval);
+        int CalculateDelay(IDictionary<string, long> taskQueue, List<TaskToWorker> taskToWorkerList, int baseSleepInterval, int currentSleepInterval);
     }
 }
